permissions in store; permissions refactor

+++ src/app/core/rest/permission.service.ts
@@ -18,15 +18,11 @@
  */
 
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpErrorResponse, HttpResponse} from '@angular/common/http';
+import {HttpClient, HttpResponse} from '@angular/common/http';
 import {Store} from '@ngrx/store';
 
-import {Permissions} from '../dto/permissions';
-import {Permission} from '../dto/permission';
+import {Permissions, Permission} from '../dto';
 import {Observable} from 'rxjs/Observable';
-import {LumeerError} from '../error/lumeer.error';
-import {ErrorObservable} from 'rxjs/observable/ErrorObservable';
-import {catchError} from 'rxjs/operators';
 import {Workspace} from '../store/navigation/workspace.model';
 import {AppState} from '../store/app.state';
 import {selectWorkspace} from '../store/navigation/navigation.state';
@@ -42,36 +38,26 @@ export abstract class PermissionService {
   }
 
   public getPermissions(): Observable<Permissions> {
-    return this.httpClient.get<Permissions>(`${this.actualApiPrefix()}/permissions`)
-      .pipe(catchError(PermissionService.handleGlobalError));
+    return this.httpClient.get<Permissions>(`${this.actualApiPrefix()}/permissions`);
   }
 
   public updateUserPermission(userPermissions: Permission): Observable<Permission> {
-    return this.httpClient.put<Permission>(`${this.actualApiPrefix()}/permissions/users`, userPermissions)
-      .pipe(catchError(PermissionService.handleGlobalError));
+    return this.httpClient.put<Permission>(`${this.actualApiPrefix()}/permissions/users`, userPermissions);
   }
 
   public updateGroupPermission(userPermissions: Permission): Observable<Permission> {
-    return this.httpClient.put<Permission>(`${this.actualApiPrefix()}/permissions/groups`, userPermissions)
-      .pipe(catchError(PermissionService.handleGlobalError));
+    return this.httpClient.put<Permission>(`${this.actualApiPrefix()}/permissions/groups`, userPermissions);
   }
 
   public removeUserPermission(user: string): Observable<HttpResponse<any>> {
-    return this.httpClient.delete(
-      `${this.actualApiPrefix()}/permissions/users/${user}`,
-      {observe: 'response', responseType: 'text'}
-    ).pipe(catchError(PermissionService.handleGlobalError));
+    return this.httpClient.delete(`${this.actualApiPrefix()}/permissions/users/${user}`,
+      {observe: 'response', responseType: 'text'})
   }
 
   public removeGroupPermission(group: string): Observable<HttpResponse<any>> {
-    return this.httpClient.delete(
-      `${this.actualApiPrefix()}/permissions/groups/${group}`,
+    return this.httpClient.delete(`${this.actualApiPrefix()}/permissions/groups/${group}`,
       {observe: 'response', responseType: 'text'}
-    ).pipe(catchError(PermissionService.handleGlobalError));
-  }
-
-  protected static handleGlobalError(error: HttpErrorResponse): ErrorObservable {
-    throw new LumeerError(error.message);
+    );
   }
 
   protected abstract actualApiPrefix(): string;
